♻️ Command bar: retire the old search/command/voice surfaces
Consolidate everything onto UnifiedCommandBar. Delete the now-dead
standalone windows — OmniSearchWindow/View and CommandBarWindow/View —
and slim VoiceCommandWindow.swift down to just the pieces the unified
bar composes, renamed to Voice/VoiceComponents.swift (VoiceCommandState,
ResultItem, TranscriptEntry, WaveBar, ListeningTimer, ResultRow). The
3-column voice window/view and DragDivider are gone.

Repoint the last callers at UnifiedCommandBarWindow: HUDTopBar's Search
button, ScreenMapWindowController's overlap-avoidance, and collapse the
AppDelegate omniSearch+commandBar surface registrations into one
"unifiedCommandBar".
